Modify the method for getting output index of metagraph.

+++ mindspore/lite/src/runtime/kernel/arm/fp32/reduce.cc
@@ -71,7 +71,7 @@ int ReduceCPUKernel::CheckParameters() {
     return RET_ERROR;
   }
   for (auto i = 0; i < num_axes_; i++) {
-    if (axes_[i] < -static_cast<int>(input_rank) || static_cast<size_t>(axes_[i]) >= input_rank) {
+    if (axes_[i] < -static_cast<int>(input_rank) || axes_[i] >= static_cast<int>(input_rank)) {
       MS_LOG(ERROR) << "Reduce got invalid axis " << axes_[i] << ", axis should be in ["
                     << -static_cast<int>(input_rank) << ", " << input_rank - 1 << "].";
       return RET_ERROR;
